flytting av utils
